refactor: useInput errorMsg 파생 상태로 관리
